+++ CHANGES.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.4.0] - (2026-05-17)
+* Add caching
+* Vectorize radiative forcing calculations
+
 ## [1.3.1] - (2026-01-29)
 * Fixed packaging issue